+++ README.md
@@ -1,26 +1,29 @@
-# laboratori_sandbox
+# test
 
 ## Project setup
 ```
-yarn install
+npm install
 ```
 
 ### Compiles and hot-reloads for development
 ```
-yarn run serve
+npm run serve
 ```
 
 ### Compiles and minifies for production
 ```
-yarn run build
+npm run build
+```
+
+### Run your tests
+```
+npm run test
 ```
 
 ### Lints and fixes files
 ```
-yarn run lint
+npm run lint
 ```
 
-### Run your unit tests
-```
-yarn run test:unit
-```
+### Customize configuration
+See [Configuration Reference](https://cli.vuejs.org/config/).
