+++ public/index.html
@@ -9,7 +9,7 @@
     content="width=device-width,initial-scale=1.0">
   <link rel="icon"
     href="<%= BASE_URL %>favicon.ico">
-  <title>laboratori_sandbox</title>
+  <title>Test</title>
   <link rel="stylesheet"
     href="https://use.fontawesome.com/releases/v5.6.1/css/all.css"
     integrity="sha384-gfdkjb5BdAXd+lj+gudLWI+BXq4IuLW5IT+brZEZsLFm++aCMlF1V92rMkPaX4PP"
@@ -20,8 +20,7 @@
 
 <body>
   <noscript>
-    <strong>We're sorry but laboratori_sandbox doesn't work properly without JavaScript enabled. Please enable it to
-      continue.</strong>
+    <strong>We're sorry but test doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
   </noscript>
   <div id="app"></div>
   <!-- built files will be auto injected -->
