+++ src/App.vue
@@ -1,33 +1,28 @@
 <template>
   <div id="app">
-    <SomeComponent msg="Welcome to Your Vue.js App"/>
-    <SearchInput/>
-    <SideBar/>
+    <img alt="Vue logo" src="./assets/logo.png">
+    <HelloWorld msg="Welcome to Your Vue.js App"/>
   </div>
 </template>
 
 <script>
-import SomeComponent from "./components/SomeComponent.vue";
-import SearchInput from "./components/SearchInput.vue";
-import SideBar from "./components/SideBar.vue"
+import HelloWorld from './components/HelloWorld.vue'
+
 export default {
-  name: "app",
+  name: 'app',
   components: {
-    SomeComponent,
-    SearchInput,
-    SideBar
+    HelloWorld
   }
-};
+}
 </script>
 
 <style>
 #app {
-  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  font-family: 'Avenir', Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
   color: #2c3e50;
   margin-top: 60px;
-  padding: 0 500px;
 }
 </style>
