+++ src/main.js
@@ -1,8 +1,13 @@
-import Vue from "vue";
-import App from "./App.vue";
+import Vue from 'vue'
+import axios from 'axios'
+import VueAxios from 'vue-axios'
 
-Vue.config.productionTip = false;
+import GettingData from './components/GettingData.vue'
+
+Vue.use(VueAxios, axios);
+
+Vue.config.productionTip = false
 
 new Vue({
-  render: h => h(App)
-}).$mount("#app");
+  render: h => h(GettingData),
+}).$mount('#app')
