--- tests/public/index.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible"
-    content="IE=edge">
-  <meta name="viewport"
-    content="width=device-width,initial-scale=1.0">
-  <link rel="icon"
-    href="<%= BASE_URL %>favicon.ico">
-  <title>Test</title>
-  <link rel="stylesheet"
-    href="https://use.fontawesome.com/releases/v5.6.1/css/all.css"
-    integrity="sha384-gfdkjb5BdAXd+lj+gudLWI+BXq4IuLW5IT+brZEZsLFm++aCMlF1V92rMkPaX4PP"
-    crossorigin="anonymous">
-  <link rel="stylesheet"
-    href="https://new-gx-ds.herokuapp.com/css/main_4.0.0.min.css">
-</head>
-
-<body>
-  <noscript>
-    <strong>We're sorry but test doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
-  </noscript>
-  <div id="app"></div>
-  <!-- built files will be auto injected -->
-</body>
-
-</html>
